services/google-mcp/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
groq
pydantic
pydantic-settings
openai
tenacity
httpx
